# Conflicts:
#	.idea/jarRepositories.xml
#	.idea/runConfigurations/Aper1_Calibration.xml
#	.idea/runConfigurations/Aper1_Myo.xml
#	.idea/runConfigurations/Aper1_R1_Vein.xml
#	.idea/runConfigurations/Aper1_R2_10_mm_PureLogic_0_5_Hz.xml
#	.idea/runConfigurations/Aper1_R2_6_mm_PureLogic_2_Hz.xml
#	.idea/runConfigurations/Aper1_R2_7_mm_PureLogic_2_Hz.xml
#	.idea/runConfigurations/Aper1_R2_8_mm_PureLogic_2_Hz.xml
#	.idea/runConfigurations/Aper2_ECG.xml
#	.idea/runConfigurations/Briko.xml
#	.idea/runConfigurations/ConverterAper.xml
#	.idea/runConfigurations/Kleiber_Myo.xml
#	.idea/runConfigurations/Loopback.xml
#	.idea/runConfigurations/NmisRsce.xml
#	.idea/runConfigurations/Prv.xml
#	.idea/runConfigurations/Rcm.xml
#	.idea/runConfigurations/RcmCalibration.xml
#	.idea/runConfigurations/SunTechNIBPTest_Aper1.xml
#	.idea/runConfigurations/SunTechNIBP_Aper1.xml
#	.idea/runConfigurations/SunTechNIBP_Aper2.xml
#	Communication/src/test/java/com/ak/appliance/aper/comm/converter/AperStage4Current2VariableTest.java
#	Desktop/src/main/java/com/ak/comm/converter/prv/PrvVariable.java
#	Desktop/src/main/java/com/ak/spring/SpringFxApplication.java
#	Desktop/src/main/resources/application.yml
#	build.gradle

+++ Util/src/main/java/com/ak/util/Clean.java
@@ -7,41 +7,40 @@ import java.nio.file.Files;
 import java.nio.file.LinkOption;
 import java.nio.file.Path;
 import java.util.Arrays;
+import java.util.Objects;
 import java.util.logging.Level;
 import java.util.logging.Logger;
-
-import javax.annotation.Nonnull;
 
 public enum Clean {
   ;
 
-  public static void clean(@Nonnull Cleaner.Cleanable[] toClean) {
+  private static final Logger LOGGER = Logger.getLogger(Clean.class.getName());
+
+  public static void clean(Cleaner.Cleanable... toClean) {
     Arrays.stream(toClean).forEach(Cleaner.Cleanable::clean);
   }
 
-  public static void clean(@Nonnull Path root) {
-    Logger.getLogger(Clean.class.getName()).log(Level.INFO, () -> "Clean directory %s".formatted(root));
+  public static void clean(Path root) {
+    LOGGER.log(Level.INFO, () -> "Clean directory %s".formatted(root.toAbsolutePath()));
     cleanRecursive(root);
     try {
       Files.deleteIfExists(root);
     }
     catch (IOException e) {
-      Logger.getLogger(Clean.class.getName()).log(Level.WARNING, root.toString(), e.getMessage());
+      LOGGER.log(Level.WARNING, root.toString(), e);
     }
   }
 
-  private static void cleanRecursive(@Nonnull Path path) {
-    if (Files.exists(path, LinkOption.NOFOLLOW_LINKS)) {
+  private static void cleanRecursive(Path path) {
+    if (Files.exists(Objects.requireNonNull(path), LinkOption.NOFOLLOW_LINKS) && Files.isDirectory(path, LinkOption.NOFOLLOW_LINKS)) {
       try (DirectoryStream<Path> ds = Files.newDirectoryStream(path)) {
-        for (Path file : ds) {
-          if (Files.isDirectory(file, LinkOption.NOFOLLOW_LINKS)) {
-            cleanRecursive(file);
-          }
+        for (Path file : Objects.requireNonNull(ds)) {
+          cleanRecursive(file);
           Files.deleteIfExists(file);
         }
       }
-      catch (IOException e) {
-        Logger.getLogger(Clean.class.getName()).log(Level.WARNING, path.toString(), e.getMessage());
+      catch (IOException | NullPointerException e) {
+        LOGGER.log(Level.WARNING, path.toString(), e);
       }
     }
   }
